requirements.txt
# Core Data Science & Analytics
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database & Data Processing
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Financial Analytics
yfinance==0.2.28
pandas-datareader==0.10.0
empyrical==0.5.5

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Machine Learning & Anomaly Detection
isolation-forest==0.1.0
pyod==1.1.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0 